requirements.txt
amqp==2.5.2
autopep8==1.4.4
Babel==2.7.0
backcall==0.1.0
billiard==3.6.1.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.0
certifi==2019.6.16
cffi==1.13.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.2
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.3
django-cors-headers==3.1.0
django-filter==2.2.0
django-phonenumber-field==3.0.1
django-silk==3.0.2
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
drf-yasg==1.16.1
gcloud==0.18.3
googleapis-common-protos==1.51.0
gprof2dot==2017.9.19
html5lib==1.0.1
httplib2==0.17.2
idna==2.8
importlib-metadata==1.3.0
inflection==0.3.1
ipython==7.7.0
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.14.1
Jinja2==2.10.1
kombu==4.6.7
MarkupSafe==1.1.1
more-itertools==8.0.2
newrelic==5.12.0.140
oauth2client==4.1.3
openapi-codec==1.3.2
packaging==19.0
parso==0.5.1
pexpect==4.7.0
phonenumbers==8.10.17
phonenumberslite==8.10.15
pickleshare==0.7.5
Pillow==6.1.0
prompt-toolkit==2.0.9
protobuf==3.11.3
psycopg2==2.8.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pyparsing==2.4.1.1
Pyphen==0.9.5
python-dateutil==2.8.0
pytz==2019.1
redis==3.3.11
requests==2.22.0
rsa==4.0
ruamel.yaml==0.16.0
ruamel.yaml.clib==0.1.0
sendgrid
sendgrid-django==4.2.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tinycss2==1.0.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
WeasyPrint==50
webencodings==0.5.1
xlwt==1.3.0
zipp==0.6.0
